Other progress and auto-load of radiated cuboids

+++ src/main/java/com/mochibit/defcon/radiation/CuboidVertex.kt
@@ -1,7 +1,7 @@
 package com.mochibit.defcon.radiation
 
 data class CuboidVertex(
-    val x: Double,
-    val y: Double,
-    val z: Double
+    val x: Int,
+    val y: Int,
+    val z: Int,
 )
